基本完结版

+++ src/samesun/java/vod/controller/vodsession/VodSessionController.java
@@ -8,6 +8,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.apache.log4j.Logger;
+import org.hibernate.criterion.DetachedCriteria;
+import org.hibernate.criterion.Property;
+import org.hibernate.criterion.Restrictions;
 import org.jeecgframework.core.common.controller.BaseController;
 import org.jeecgframework.core.common.hibernate.qbc.CriteriaQuery;
 import org.jeecgframework.core.common.model.json.AjaxJson;
@@ -24,9 +27,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
+import vod.entity.confrtspsrvinfo.ConfRtspSrvInfoEntity;
+import vod.entity.meetinginfo.MeetingInfoEntity;
 import vod.entity.vodsectionrecord.VodSectionRecordEntity;
 import vod.entity.vodsession.VodSessionEntity;
 import vod.page.vodsession.VodSessionPage;
+import vod.samesun.util.NetTelnet;
 import vod.samesun.util.SystemType;
 import vod.service.vodsession.VodSessionServiceI;
 
@@ -70,6 +76,16 @@ public class VodSessionController extends BaseController {
 	public ModelAndView vodSession(HttpServletRequest request) {
 		return new ModelAndView("vod/vodsession/vodSessionList");
 	}
+	
+	/**
+	 * 点播信息会话列表 页面跳转
+	 * 
+	 * @return
+	 */
+	@RequestMapping(params = "vodSession4Training")
+	public ModelAndView vodSession4Training(HttpServletRequest request) {
+		return new ModelAndView("vod/vodSession4Training/vodSessionList");
+	}
 
 	/**
 	 * easyui AJAX请求数据
@@ -82,13 +98,33 @@ public class VodSessionController extends BaseController {
 
 	@SuppressWarnings("unchecked")
 	@RequestMapping(params = "datagrid")
-	public void datagrid(VodSessionEntity vodSession,HttpServletRequest request, HttpServletResponse response, DataGrid dataGrid) {
+	public void datagrid(VodSessionEntity vodSession,HttpServletRequest request, HttpServletResponse response, DataGrid dataGrid, String rightid) {
 		dataGrid.setSort("begindt");
 		dataGrid.setOrder(SortDirection.desc);
+		
 		CriteriaQuery cq = new CriteriaQuery(VodSessionEntity.class, dataGrid);
+		DetachedCriteria cqdc = cq.getDetachedCriteria();
+		cqdc.add(Restrictions.isNotNull("begindt"));
+		cqdc.add(Restrictions.isNotNull("enddt"));
+		
+		//关联会议种类
+		DetachedCriteria cq_meeting = DetachedCriteria.forClass(MeetingInfoEntity.class);
+		cq_meeting.add(Restrictions.isNotNull("rightid"));
+		cq_meeting.add(Restrictions.eq("rightid", rightid));
+		cq_meeting.setProjection(Property.forName("id"));
+		cqdc.add(Property.forName("meetingid").in(cq_meeting));
+		
+		//关联明细状态
+		DetachedCriteria cq_vsr = DetachedCriteria.forClass(VodSectionRecordEntity.class);
+		cq_vsr.add(Restrictions.eq("recState", new Integer(SystemType.REC_STATE_5)));
+		cq_vsr.setProjection(Property.forName("sessionid"));
+		cqdc.add(Property.forName("liveSession").in(cq_vsr));
+		
+		cq.setDetachedCriteria(cqdc);
 		//查询条件组装器
 		org.jeecgframework.core.extend.hqlsearch.HqlGenerateUtil.installHql(cq, vodSession, request.getParameterMap());
 		this.vodSessionService.getDataGridReturn(cq, true);
+		
 		List<VodSessionEntity> sessions = dataGrid.getResults();
 		List<VodSessionPage> pages = new ArrayList<VodSessionPage>();
 		if(sessions != null && sessions.size() > 0){
@@ -102,16 +138,18 @@ public class VodSessionController extends BaseController {
 				}
 				Date begin = page.getBegindt();
 				Date end = page.getEnddt();
-				if(begin != null && end != null && hasdetail(e)){
-					int time = (int) ((end.getTime() - begin.getTime()) / (60 * 1000));
-					page.setDuration(time);
-					pages.add(page);
+				int time = (int) ((end.getTime() - begin.getTime()) / (60 * 1000));
+				page.setDuration(time);
+				if(SystemType.MEETING_RIGHT_1.equals(rightid)){
+					page.setTypename(systemService.getType(page.getTypeid().toString(), SystemType.MEETING_TYPE).getTypename());
+				}else if(SystemType.MEETING_RIGHT_2.equals(rightid)){
+					page.setTypename(systemService.getType(page.getTypeid().toString(), SystemType.TRAINING_TYPE).getTypename());
 				}
+				pages.add(page);
 				
 			}
 		}
 		dataGrid.setResults(pages);
-		dataGrid.setTotal(pages.size());
 		TagUtil.datagrid(response, dataGrid);
 	}
 
@@ -126,10 +164,35 @@ public class VodSessionController extends BaseController {
 		AjaxJson j = new AjaxJson();
 		vodSession = systemService.getEntity(VodSessionEntity.class, vodSession.getId());
 		message = "点播信息会话删除成功";
-		vodSessionService.delete(vodSession);
-		systemService.addLog(message, Globals.Log_Type_DEL, Globals.Log_Leavel_INFO);
-		
-		j.setMsg(message);
+		List<VodSectionRecordEntity> sections = systemService.findByProperty(VodSectionRecordEntity.class, "sessionid", vodSession.getLiveSession());
+
+		//先删除文件
+		for(VodSectionRecordEntity s : sections){
+			String strFileName = s.getFilename();
+			String RtspSrvID = s.getRtspsrvid();
+			String RtspRelativeDir = s.getRtsprelativedir();
+			ConfRtspSrvInfoEntity rtsp = systemService.get(ConfRtspSrvInfoEntity.class, RtspSrvID);
+			if(StringUtil.isNotEmpty(strFileName) && rtsp != null){
+				String strIP = rtsp.getIpaddress();
+				String strPort = "8000";//默认端口
+				message = NetTelnet.deleteFiles4Telnet(strIP, strPort, RtspRelativeDir, strFileName);
+			}
+		}
+		if (StringUtil.isNotEmpty(message) && message.indexOf("error") > 0){
+			try {
+				message = "点播信息会话删除失败";
+				throw new Exception(message);
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
+		}else{
+			//再删除明细
+			systemService.deleteAllEntitie(sections);
+			//最后删除session
+			vodSessionService.delete(vodSession);
+			systemService.addLog(message, Globals.Log_Type_DEL, Globals.Log_Leavel_INFO);
+			j.setMsg(message);
+		}
 		return j;
 	}
 
@@ -183,11 +246,12 @@ public class VodSessionController extends BaseController {
 	 * @param e
 	 * @return
 	 */
+	@SuppressWarnings("unused")
 	private boolean hasdetail(VodSessionEntity e){
 		boolean flag = false;
 		List<VodSectionRecordEntity> vods = systemService.findByProperty(VodSectionRecordEntity.class, "sessionid", e.getLiveSession());
 		for(VodSectionRecordEntity v : vods){
-			if(Integer.valueOf(SystemType.REC_STATE_5) == v.getRecState()){
+			if(SystemType.REC_STATE_5.equals(v.getRecState().toString())){
 				flag = true;
 			}
 		}
